requirements.txt
# ── Core ML & Deep Learning ──────────────────────────────────────────────────
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
# PyTorch Geometric (install via separate pip command — see README)
torch_geometric==2.5.3
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-cluster==1.6.3
torch-spline-conv==1.2.2

# ── Tabular / Gradient Boosting ───────────────────────────────────────────────
xgboost==2.0.3
scikit-learn==1.4.2
imbalanced-learn==0.12.2

# ── Graph / Network ───────────────────────────────────────────────────────────
networkx==3.3
pyvis==0.3.2

# ── Explainability ────────────────────────────────────────────────────────────
shap==0.45.1

# ── LLM / API ─────────────────────────────────────────────────────────────────
openai==1.30.1
transformers==4.41.2
accelerate==0.30.1
sentencepiece==0.2.0

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit==1.35.0
plotly==5.22.0

# ── Data / Numerics ───────────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
pyarrow==16.1.0

# ── Config / Utilities ────────────────────────────────────────────────────────
pyyaml==6.0.1
omegaconf==2.3.0
tqdm==4.66.4
python-dotenv==1.0.1
rich==13.7.1

# ── Evaluation / Metrics ─────────────────────────────────────────────────────
matplotlib==3.9.0
seaborn==0.13.2
